Changing root to rootCA
