+++ orchestrator-bundle/orc8r-certifier-operator/tests/unit/lib/charms/magma_orc8r_certifier/v0/test_cert_root_ca_requirer.py
@@ -5,7 +5,7 @@ import unittest
 from unittest.mock import PropertyMock, patch
 
 from ops import testing
-from test_charms.test_root_requirer.src.charm import (  # type: ignore[import]
+from test_charms.test_root_ca_requirer.src.charm import (  # type: ignore[import]
     WhateverCharm,
 )
 
@@ -18,10 +18,10 @@ class TestCertRootRequires(unittest.TestCase):
         self.harness = testing.Harness(WhateverCharm)
         self.addCleanup(self.harness.cleanup)
         self.harness.begin()
-        self.relationship_name = "cert-root"
+        self.relationship_name = "cert-root-ca"
 
     @patch(
-        "test_charms.test_root_requirer.src.charm.WhateverCharm.CERT_PATH",
+        "test_charms.test_root_ca_requirer.src.charm.WhateverCharm.CERT_PATH",
         new_callable=PropertyMock,
     )
     @patch("ops.model.Container.push")
