+++ orchestrator-bundle/orc8r-certifier-operator/tests/unit/lib/charms/magma_orc8r_certifier/v0/test_charms/test_root_ca_requirer/metadata.yaml
@@ -14,5 +14,5 @@ resources:
     description: whatever image
 
 requires:
-  cert-root:
-    interface: cert-root
+  cert-root-ca:
+    interface: cert-root-ca
